Creating a super-simple ascii server
A very simple ASCII server that renders the text visualization in the
browser.

Cleaned up the ascii server

Added a real websocket protocol, fixed the ugly template path hack, and
implemented resetting.

Minor tweak

Added some documentation

add tornado

Style cleanup

Cleaning up some of the code following @kfuruya review.

Fixed out-of-range step bug

Fixing issue #33, bug which caused the server to crash when an
out-of-range step was requested.

Implementing Play button

Adding a play button, which runs the model at the set frame rate until
it hits the end.

Tabs to spaces

Bringing it in line with the master...